Exceptions

Module

PDFHighlights.Internal.Exceptions — Module

This module contains all the exceptions used in the package, as well as the macro(s) that they were created by.

Types

PDFHighlights.Internal.Exceptions.DirectoryDoesNotExist — Type

DirectoryDoesNotExist <: Exception

Exception thrown when the specified directory doesn't exist.

Fields

file::String: absolute or relative path to a file that is associated with the exception

See also: @exception_with_file

PDFHighlights.Internal.Exceptions.DoesNotExist — Type

DoesNotExist <: Exception

Exception thrown when the specified file or directory doesn't exist.

Fields

file::String: absolute or relative path to a file that is associated with the exception

See also: @exception_with_file

PDFHighlights.Internal.Exceptions.FileDoesNotExist — Type

FileDoesNotExist <: Exception

Exception thrown when the specified file doesn't exist.

Fields

file::String: absolute or relative path to a file that is associated with the exception

See also: @exception_with_file

PDFHighlights.Internal.Exceptions.IncorrectHeader — Type

IncorrectHeader <: Exception

Exception thrown when the specified file has an incorrect header.

Fields

file::String: absolute or relative path to a file that is associated with the exception

See also: @exception_with_file

PDFHighlights.Internal.Exceptions.IntegrityCheckFailed — Type

IntegrityCheckFailed <: Exception

An exception thrown when another exception was thrown while checking the integrity of the table.

Fields

file::String: absolute or relative path to a file that is associated with the exception

See also: @exception_with_file

PDFHighlights.Internal.Exceptions.LastElementIsNotAnInteger — Type

LastElementIsNotAnInteger <: Exception

Exception thrown when the last element in the line is not an integer.

Fields

file::String: absolute or relative path to a file that is associated with the exception

line::Int: line number of the specified file to which this exception is associated

See also: @exception_with_line

PDFHighlights.Internal.Exceptions.NotCSV — Type

NotCSV <: Exception

Exception thrown when the specified path does not end in .csv.

Fields

file::String: absolute or relative path to a file that is associated with the exception

See also: @exception_with_file

PDFHighlights.Internal.Exceptions.NotCSVorDir — Type

NotCSVorDir <: Exception

An exception thrown when the specified path does not end in .csv ans is not a directory.

Fields

file::String: absolute or relative path to a file that is associated with the exception

See also: @exception_with_file

PDFHighlights.Internal.Exceptions.NotCSVorPDForDir — Type

NotCSVorPDForDir <: Exception

An exception thrown when the specified path does not end in .csv or .pdf and is not a directory.

Fields

file::String: absolute or relative path to a file that is associated with the exception

See also: @exception_with_file

PDFHighlights.Internal.Exceptions.NotPDF — Type

NotPDF <: Exception

Exception thrown when the specified path does not end in .pdf.

Fields

file::String: absolute or relative path to a file that is associated with the exception

See also: @exception_with_file

PDFHighlights.Internal.Exceptions.NotSixColumns — Type

NotSixColumns <: Exception

Exception thrown when the row does not represent elements for 6 columns.

Fields

file::String: absolute or relative path to a file that is associated with the exception

line::Int: line number of the specified file to which this exception is associated

See also: @exception_with_line

PDFHighlights.Internal.Exceptions.OnlyOneContext — Type

OnlyOneContext <: Exception

Exception thrown when more than one context has been passed.

See also: @exception_without_fields

PDFHighlights.Internal.Exceptions.OnlyOneEquation — Type

OnlyOneEquation <: Exception

Exception thrown when non-context equation has been passed.

See also: @exception_without_fields

PDFHighlights.Internal.Exceptions.SymbolIsNotSupported — Type

SymbolIsNotSupported <: Exception

Exception thrown when the specified symbol is not supported.

Fields

symbol::Symbol: the symbol associated with the exception

See also: @exception_with_symbol

Macros

PDFHighlights.Internal.Exceptions.@docstring — Macro

@docstring(macro_name::Symbol) -> Expr

Redefine the docstring variable in the innermost macro.

Arguments

macro_name::Symbol: name of the macro that generated the exception

Returns

Expr: docstring redefinition

Example

using PDFHighlights
using SyntaxTree

macro_name = QuoteNode(:name)

@macroexpand(PDFHighlights.Internal.Exceptions.@docstring(name)) |> linefilter! ==
quote
    macro_name = $(macro_name)
    docstring = Meta.quot(
        """
            $(exception_name) <: Exception

        $(docstring)

        $(fields)

        See also: [`@$(macro_name)`](@ref)
        """
    )
end |> linefilter!

See also: @exception

PDFHighlights.Internal.Exceptions.@exception — Macro

@exception(macro_name::Symbol, args::Expr...; context::Expr=:()) -> Expr

Create a macro to create exceptions. Optionally inject context before defining the structure.

Arguments

macro_name::Symbol: name of the macro

args::Tuple{Vararg{Expr}}: a set of fields to be defined in the exception structure

Keywords

context::Expr=:(): expression evaluated before defining the exception structure

Returns

Expr: new macro definition

Throws

OnlyOneContext: more than one context has been passed

OnlyOneEquation: more than one equation has been passed

Example

using PDFHighlights
using Suppressor
using SyntaxTree

macro_name = :name
args = (:(arg1::String), :(arg2::Int))
context = :()

d1 = @capture_out @macroexpand(
    PDFHighlights.Internal.Exceptions.@exception(
        name,
        arg1::String,
        arg2::Int,
    )
) |> linefilter! |> dump

d2 = @capture_out quote
    macro $(macro_name)(
        exception_name::Symbol,
        docstring::Union{Expr, String},
        error_message_bits::Union{Expr, String}...,
    )
        module_name = __module__

        args = $(args)
        error_header = "$(module_name).$(exception_name):"

        $(context)

        return esc(
            quote
                @doc $(docstring)
                mutable struct $(exception_name) <: Exception
                    $(args...)
                    $(exception_name)($(args...)) = new($(args...))
                end

                Base.showerror(io::IO, e::$(module_name).$(exception_name)) =
                print(
                    io, string(
                        '\n', '\n',
                        $(error_header), '\n',
                        $(error_message_bits...), '\n',
                    )
                )
            end
        )
    end
end |> linefilter! |> dump

d1 == d2

PDFHighlights.Internal.Exceptions.@exception_with_file — Macro

@exception_with_file(
    exception_name::Symbol,
    docstring::Union{Expr, String},
    error_message_bits::Tuple{Vararg{Union{Expr, String}}},
) -> Expr

Create an exception with one field: a path to the file.

Arguments

exception_name::Symbol: name of the exception

docstring::Union{Expr, String}: documentation string

error_message_bits::Tuple{Vararg{Union{Expr, String}}}: strings and expressions which will be interpolated in the showerror output

Returns

Expr: an exception definition (struct + showerror overload)

See also: @exception

PDFHighlights.Internal.Exceptions.@exception_with_line — Macro

@exception_with_line(
    exception_name::Symbol,
    docstring::Union{Expr, String},
    error_message_bits::Tuple{Vararg{Union{Expr, String}}},
) -> Expr

Create an exception with two fields: a path to the file and a line number.

Arguments

exception_name::Symbol: name of the exception

docstring::Union{Expr, String}: documentation string

error_message_bits::Tuple{Vararg{Union{Expr, String}}}: strings and expressions which will be interpolated in the showerror output

Returns

Expr: an exception definition (struct + showerror overload)

See also: @exception

PDFHighlights.Internal.Exceptions.@exception_with_symbol — Macro

@exception_with_symbol(
    exception_name::Symbol,
    docstring::Union{Expr, String},
    error_message_bits::Tuple{Vararg{Union{Expr, String}}},
) -> Expr

Create an exception with one field: a symbol.

Arguments

exception_name::Symbol: name of the exception

docstring::Union{Expr, String}: documentation string

error_message_bits::Tuple{Vararg{Union{Expr, String}}}: strings and expressions which will be interpolated in the showerror output

Returns

Expr: an exception definition (struct + showerror overload)

See also: @exception

PDFHighlights.Internal.Exceptions.@exception_without_fields — Macro

@exception_without_fields(
    exception_name::Symbol,
    docstring::Union{Expr, String},
    error_message_bits::Tuple{Vararg{Union{Expr, String}}},
) -> Expr

Create an exception with no fields.

Arguments

exception_name::Symbol: name of the exception

docstring::Union{Expr, String}: documentation string

error_message_bits::Tuple{Vararg{Union{Expr, String}}}: strings and expressions which will be interpolated in the showerror output

Returns

Expr: an exception definition (struct + showerror overload)

See also: @exception

PDFHighlights.Internal.Exceptions.@fields — Macro

@fields(strings::String...) -> Expr

Define the fields variable in the innermost macro. This variable is a string that describes type fields.

Arguments

strings::Tuple{Vararg{String}}: strings to store in the fields variable

Returns

Expr: fields definition

Example

using PDFHighlights
using SyntaxTree

strings = ("`Cloverfield`: there's no place like home",)

@macroexpand(
    PDFHighlights.Internal.Exceptions.@fields(
        "`Cloverfield`: there's no place like home"
    )
) |> linefilter! == quote
    strings = $(strings) |> collect
    strings = join(["- " * string for string in strings], '\n')
    fields = """
    # Fields\n
    $(strings)
    """
end |> linefilter!

See also: @exception